Add base projects
Base projects are responsible of providing the running host EXE and the
implemented/exposed host capabilities available to all plugins.

The host Application should be responsible of implementing the loading
means for the up coming plugins.

The Core package is responsible of exposing and defining all of the
host and plugin abstractions. So basically this project will act as a
contracts/abstractions project.