Add WPF Base Host Application Components
The main application consists of a main window and the infrastructure
required for that window to resolve the remaining types or view models
in this case.

In order to provide a better developer experience a dynamic object is
used for locating the viewmodels and for accessing the IoC Container
instance.

The BaseServicesModule is an Autofac module that registers both logging
and configuration implementations.

Since the OnStartup override in the App class allows the application to
get the commandline arguments, which in turn can be used for initialize
the configuration implementation in the BaseServicesModule.

Also is worth mentioning that the App class is the default root
ViewModel Locator.
The reason for this is that the class is the only class in the
application that has actually access to the IoC instance, so in order to
avoid referencing the IoC elsewhere for ViewModel resolution, it's used
at App Level.

The ViewModelFlyweightFactory is responsible of caching and resolving
viewmodels from any given view that has a DataContext set to

```xaml
DataContext="{Binding Source={StaticResource ViewModelFlyweightFactory},
Path=SomeVM}
```