Add main component initial implementation
The main component will expose the basics for other application
components.

The idea for the main component is to have a work area that can be
shaped by other components by adding pieces or UI components via
DataTemplates.

Registrations for commands and view models are carried out automatically
on the Wpf Autofac module.
This module is responsible of performing a post initialization step that
the ViewModelLocator implementation requires right after the IoC has
been initialized.
Perhaps if another IoC is used this can still be the case or further
changes would be required. Either way I don't want to make this very
Autofac specific at all.