Add plugin loading infrastructure
The plugin loading infrastrucutre is based on AssemblyLoadContext.
This lets plugin assemblies to host dependencies that are not referenced
in the main application or use dependencies whose version is different
from the one distributed with the application.

This class makes heave use of Autofac AssemblyContext feature, although
it's still possible to use another IoC implementation