Add plugin sources
Plugin sources allow different abstractions for locating and providing
plugins to the application.

In this very first version, File based and composite sources are
provided.

The composite is a MUST since many sources can be attached and this
implementation hides that fact also provides a yet simple way to extend
and use other sources.

The file based is aimed for json manifests. But there's still the
possibility to detect other DLLs by inspecting their contents.