Add Plugin composition module
This module brings and registers all of the plugin functionality into
the application, no further configuration is needed.

The plugin registry is the first class that aims for detecting new
plugins and act as a source for the UI as well as acting as a collection
for all known to the instance plugins.